libutils, kernel: add standard functions needed by OpenSSL
Prepare for building OpenSSL libcrypto.
1. Add the following to libutils:
- isdigit(), isspace(), isalpha(), isalnum(), isxdigit(), isupper(),
  islower(), toupper(), tolower(): basic ASCII-only implementation
- atoi(), sscanf() (stub!), strtoul(), strncmp(), strcpy(), strncpy(),
  strchr(), strcat()
2. Add abort() and time() to core/kernel.

+++ lib/libutils/isoc/include/unistd.h
@@ -31,6 +31,6 @@
 #include <stddef.h>
 
 #define __ssize_t_defined
-typedef int32_t ssize_t;
+typedef intptr_t ssize_t;
 
 #endif
